+++ crates/rpc/src/lib.rs
@@ -1,5 +1,9 @@
+mod client;
 mod server;
+#[cfg(test)]
+mod tests;
 
+pub use client::{ApiClient, MetadataInterceptor};
 pub use server::Rpc;
 
 // CONSTANTS